+++ requirements.txt
@@ -1,1 +1,1 @@
-molo.core<4.0,>=3.0
+molo.core==3.0rc7
